cleanup(virtctl): Merge ssh and scp functional tests
Merge the functional tests for virtctl ssh and scp into a single test
suite so they can share the same virtual machine.
