+++ tests/virtctl/BUILD.bazel
@@ -9,8 +9,7 @@ go_library(
         "guestfs.go",
         "imageupload.go",
         "memorydump.go",
-        "scp.go",
-        "ssh.go",
+        "ssh_scp.go",
         "usbredir.go",
         "vnc.go",
     ],
